# swagger.yml
swagger: '2.0'
info:
  description: This is a Booking Ticket API
  version: 1.0.0
  title: Booking Ticket API
  # put the contact info for your development or API team
  contact:
    email: alfinasution64@gmail.com

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used for organizing operations
...
paths:
...
  /booking/user/{userId}:
    get:
      tags:
      - schedule
      summary: List User booking
      operationId: listUserBooking
      description: |
        By accessing this, will show list of user booking in booking ticket
      produces:
      - application/json      
      responses:
        200:
          description: success get list of user booking in booking ticket
          schema:
            type: object    
            properties:      
              meta:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "successfull"
              data:                                                                        
                $ref: '#/definitions/responseBookingUser'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/internalServerErrArrayResponse'
definitions:
...
  internalServerErrArrayResponse:
    type: object    
    properties:      
      meta:
        type: object
        properties:
          status:
            type: integer
            example: 500
          message:
            type: string
            example: "Internal Server Error"
      data:
        type: array
        items:
          type: object
...
  responseBookingUser:
    type: object    
    properties:
      id:
        type: integer
        example: 1      
      titileMovie:
        type: string
        example: 'Shangchi'
      posterMovie:
        type: string
        example: 'https://dummyimage.com/400x600/000/fff'
      name:
        type: string
        example: 'Puri Indah XXI'
      timeSchedule:
        type: string
        example: '2020-02-02 18:20:20'      
...
host: virtserver.swaggerhub.com
basePath: /alfi2811/booking-ticket/1.0.0
schemes:
 - https